Enhance reliability
After a reboot to Windows, which had applied a different configuration, I wasn't
successful in reapplying the old settings from within Linux in a reliable way so
that they would be loaded by the device on the next reconnect.  Writing the
configuration twice in a row seemed to help.  This also seems to work.  There
might be some timing or what not behind the issue, I don't know.

The GUI tool is a bit prone to quitting on LIBUSB_ERROR_PIPE on my system at
the moment.  Not very user-friendly.  But I'm tired of inspecting it already,
unplugging and re-plugging in the mouse all the time like an idiot...